+++ R/AnnData.R
@@ -158,6 +158,7 @@ AnnData <- function(
   )
 }
 
+#' @importFrom methods as
 .check_matrix <- function(X) {
   if (is(X, "sparseMatrix") && !is(X, "dgCMatrix") && !is(X, "dgRMatrix")) {
     X <- as(X, "CsparseMatrix")
@@ -678,7 +679,11 @@ AnnDataR6 <- R6::R6Class(
     #' @field X Data matrix of shape `n_obs` × `n_vars`.
     X = function(value) {
       if (missing(value)) {
-        py_to_r(private$.anndata$X)
+        out <- py_to_r(private$.anndata$X)
+        if (!is.null(out)) {
+          dimnames(out) <- dimnames(self)
+        }
+        out
       } else {
         value <- .check_matrix(value)
         private$.anndata$X <- value
@@ -738,7 +743,7 @@ AnnDataR6 <- R6::R6Class(
         py_to_r(private$.anndata$layers)
       } else {
         # add check for value
-        if (!is.null(value)) {
+        if (!is.null(value) && is.list(value)) {
           for (i in seq_along(value)) {
             value[[i]] <- .check_matrix(value[[i]])
           }
@@ -772,8 +777,11 @@ AnnDataR6 <- R6::R6Class(
       if (missing(value)) {
         py_to_r(private$.anndata$obs)
       } else {
-        # add check for value
-        private$.anndata$obs <- value
+        if (is.null(value)) {
+          py_del_attr(private$.anndata, "obs")
+        } else {
+          private$.anndata$obs <- value
+        }
         self
       }
     },
@@ -784,6 +792,7 @@ AnnDataR6 <- R6::R6Class(
       } else {
         # add check for value
         private$.anndata$obs_names <- value
+        private$.anndata$obs_names$name <- attr(value, "name")
         self
       }
     },
@@ -820,8 +829,11 @@ AnnDataR6 <- R6::R6Class(
       if (missing(value)) {
         py_to_r(private$.anndata$var)
       } else {
-        # add check for value
-        private$.anndata$var <- value
+        if (is.null(value)) {
+          py_del_attr(private$.anndata, "var")
+        } else {
+          private$.anndata$var <- value
+        }
         self
       }
     },
@@ -832,6 +844,7 @@ AnnDataR6 <- R6::R6Class(
       } else {
         # add check for value
         private$.anndata$var_names <- value
+        private$.anndata$var_names$name <- attr(value, "name")
         self
       }
     },
@@ -868,8 +881,11 @@ AnnDataR6 <- R6::R6Class(
       if (missing(value)) {
         py_to_r(private$.anndata$uns)
       } else {
-        # add check for value
-        private$.anndata$uns <- value
+        if (is.null(value)) {
+          py_del_attr(private$.anndata, "uns")
+        } else {
+          private$.anndata$uns <- value
+        }
         self
       }
     },
@@ -922,7 +938,6 @@ AnnDataR6 <- R6::R6Class(
 #'
 #' @param x An AnnData object.
 #' @param layer An AnnData layer. If `NULL`, will use `ad$X`, otherwise `ad$layers[layer]`.
-#' @param value Replacement value.
 #' @param convert Not used.
 #' @param row.names Not used.
 #' @param optional Not used.
@@ -1010,24 +1025,33 @@ py_to_r.anndata._core.anndata.AnnData <- function(x) {
   ad
 }
 
-#' @rdname AnnDataHelpers
-#' @export
-`[.AnnDataR6` <- function(x, ..., layer = NULL) {
-  as.matrix.AnnDataR6(x, layer = layer)[...]
+.process_index <- function(idx, len) {
+  if (missing(idx)) {
+    builtins <- reticulate::import_builtins(convert = FALSE)
+    idx <- builtins$slice(builtins$None)
+  } else if (is.numeric(idx)) {
+    if (any(idx <= 0)) {
+      if (!all(idx < 0)) {
+        stop("integer indices should be all positive or all negative")
+      }
+      idx <- seq_len(len)[idx]
+    }
+    idx <- as.integer(idx - 1)
+  }
+
+  idx
 }
 
 #' @rdname AnnDataHelpers
+#' @importFrom reticulate tuple
+#' @param oidx Observation indices
+#' @param vidx Variable indices
 #' @export
-`[<-.AnnDataR6` <-  function(x, ..., layer = NULL, value) {
-  mat <- as.matrix(x, layer = layer)
-  mat[...] <- value
-  dimnames(mat) <- NULL
-  if (is.null(layer)) {
-    x$X <- mat
-  } else {
-    x$layers[layer] <- mat
-  }
-  x
+`[.AnnDataR6` <- function(x, oidx, vidx) {
+  oidx <- .process_index(oidx, nrow(x))
+  vidx <- .process_index(vidx, ncol(x))
+  tup <- reticulate::tuple(oidx, vidx)
+  py_to_r(x$.__enclos_env__$private$.anndata$`__getitem__`(tup))
 }
 
 #' @rdname AnnDataHelpers
@@ -1040,10 +1064,8 @@ t.AnnDataR6 <- function(x) {
 # https://github.com/theislab/anndata/blob/58886f09b2e387c6389a2de20ed0bc7d20d1b843/anndata/tests/helpers.py#L352
 #' Test if two AnnDataR6 objects are equal
 #' @inheritParams base::all.equal
-#' @param exact Whether comparisons should be exact or not. This has a somewhat flexible
-#'   meaning and should probably get refined in the future.
 #' @export
-all.equal.AnnDataR6 <- function(target, current, exact = TRUE) {
+all.equal.AnnDataR6 <- function(target, current) {
   a <- target
   b <- current
 
@@ -1078,13 +1100,7 @@ all.equal.AnnDataR6 <- function(target, current, exact = TRUE) {
 
   match <-
     aecheck(a$obs_names, b$obs_names, "obs_names") %&%
-    aecheck(a$var_names, b$var_names, "var_names")
-
-  if (isTRUE(match) && !exact) {
-    # TODO: implement not exact.. but what does it do?
-  }
-
-  match <- match %&%
+    aecheck(a$var_names, b$var_names, "var_names") %&%
     aecheck(a$obs, b$obs, "obs") %&%
     aecheck(a$var, b$var, "var") %&%
     aecheck(a$X, b$X, "X") %&%