+++ tests/testthat/helper-expect_not_equal.R
@@ -1,0 +1,3 @@
+expect_not_equal <- function(object, expected, ...) {
+  expect_false(isTRUE(all.equal(object, expected, ...)))
+}