+++ NEWS.md
@@ -1,6 +1,8 @@
 # anndata 0.7.5.1
 
-* MINOR CHANGE: Add wrapper class for Raw objects.
+* MINOR CHANGE: Add wrapper classes for Raw and Layers objects.
+
+* MAJOR CHANGE: Calling `ad[..., ...]` now correctly returns a view of `ad` instead of returning a matrix.
 
 # anndata 0.7.5 (2020-11-19)