+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * MAJOR CHANGE: Calling `ad[..., ...]` now correctly returns a view of `ad` instead of returning a matrix.
 
+* TESTING: Extend tests based on `theislab/anndata` repository.
+
 # anndata 0.7.5 (2020-11-19)
 
 * MINOR CHANGES: Updated Python requirements to anndata 0.7.5.